requirements.txt
Django
mysqlclient==2.2.4
fpdf==1.7.2
matplotlib==3.9.2
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
pydataset==0.2.0
pyzbar==0.1.9
qrcode==7.4.2
qreader==3.14
scikit_learn==1.5.2
seaborn==0.13.2
